+++ frontend/src/routes/admin/users/AdminRegisterUserModal.tsx
@@ -150,71 +150,6 @@ export default function AdminRegisterUserModal({ userId }: {userId: number}) {
           </>
         );
       }}
-      {/* <Form.Field name="event">
-        <Form.Label>
-          Event
-        </Form.Label>
-        <br />
-        <Form.Control asChild>
-          <Select.Root
-            onValueChange={(value) => {
-              setSelectedEvent(Number(value));
-              setSelectedTeam(eventsMap[Number(value)]?.max_team_size === 1 ? 'new' : '');
-            }}
-            value={selectedEvent.toString()}
-            disabled={isLoading || !!error}
-            required
-          >
-            <Select.Trigger placeholder="Select an event..." className="!w-full" />
-            <Select.Content>
-              {filteredEvents?.map((event) => (
-                <Select.Item key={event.id} value={event.id.toString()}>
-                  {event.name}
-                </Select.Item>
-              ))}
-            </Select.Content>
-          </Select.Root>
-        </Form.Control>
-      </Form.Field>
-
-      {selectedEvent && eventsMap[selectedEvent]?.max_team_size > 1 && (
-        <Form.Field name="event">
-          <Form.Label>
-            Team
-          </Form.Label>
-          <br />
-          <Form.Control asChild>
-            <Select.Root onValueChange={(value) => setSelectedTeam(value)} value={selectedTeam} disabled={teamsLoading || !!teamsError}>
-              <Select.Trigger placeholder="Select a team..." className="!w-full" />
-              <Select.Content>
-                <Select.Item value="new">Create new...</Select.Item>
-                {filteredTeams?.map((team) => (
-                  <Select.Item key={team.id} value={team.invite_code!.toString()}>
-                    {team.name}
-                    {' '}
-                    (
-                    {team.member_count}
-                    /
-                    {eventsMap[selectedEvent].max_team_size}
-                    )
-                  </Select.Item>
-                ))}
-              </Select.Content>
-            </Select.Root>
-          </Form.Control>
-        </Form.Field>
-      )}
-
-      {((selectedEvent && eventsMap[selectedEvent].max_team_size === 1) || selectedTeam === 'new') && (
-        <Form.Field name="teamName">
-          <Form.Label>Team Name</Form.Label>
-          <Form.Control asChild>
-            <TextField.Root required />
-          </Form.Control>
-          <Form.Message match="valueMissing">Please provide a team name.</Form.Message>
-        </Form.Field>
-      )} */}
-
     </Modal>
   );
 }
